requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.0.0
scikit-learn>=1.4.0
numpy>=1.26.0
torch>=2.1.0
torchvision>=0.16.0
datasets>=2.18.0
openai>=1.0.0
httpx>=0.27.0
pillow>=10.0.0
